Avoid flickering of the video frame during a trajectory calculation
The flickering happened because the QImage used to paint the video frame references the data array originally allocated by OpenCV. So, when Qt refreshed the frame (on a resize event, for example) it read the data from the same buffer another thread uses for getting the frames. This was resolved by simply deep-copying the background QImage and setting it to the bgRect when the video is paused.

Of course that means that the problem still occurs if there's a calculation in place and the video is running, but now the user has the option to wait until the calculation is over and avoid any problems.